pyproject.toml
[project]
name = "HabitTracker"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiohttp>=3.11.18",
    "python-dotenv>=1.1.0",
]
